# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Meeting Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Meeting {
  id: number;
  filename: string;
  original_filename: string;
  file_size: number;
  file_format: string;
  /** uploaded | transcribing | transcribed | failed */
  status: string;
  transcript?: string | null;
  transcribed_at?: string | null;
  uploaded_at: string;
}

export type MeetingStatsFormats = {[key: string]: number};

export interface MeetingStats {
  total_meetings: number;
  total_size_bytes: number;
  formats: MeetingStatsFormats;
  recent_uploads: number;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}