# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Meeting Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: meetings
    description: Meeting operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List all meetings
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
  /meetings/stats:
    get:
      operationId: getMeetingStats
      tags: [meetings]
      summary: Get meeting statistics
      responses:
        "200":
          description: Meeting statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingStats"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a meeting by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /meetings/{id}/transcribe:
    post:
      operationId: retryTranscription
      tags: [meetings]
      summary: Trigger or retry transcription for a meeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transcription started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already transcribing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Meeting:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        original_filename:
          type: string
        file_size:
          type: integer
        file_format:
          type: string
        status:
          type: string
          description: "uploaded | transcribing | transcribed | failed"
        transcript:
          type: string
          nullable: true
        transcribed_at:
          type: string
          nullable: true
        uploaded_at:
          type: string
      required:
        - id
        - filename
        - original_filename
        - file_size
        - file_format
        - status
        - uploaded_at
    MeetingStats:
      type: object
      properties:
        total_meetings:
          type: integer
        total_size_bytes:
          type: integer
        formats:
          type: object
          additionalProperties:
            type: integer
        recent_uploads:
          type: integer
      required:
        - total_meetings
        - total_size_bytes
        - formats
        - recent_uploads
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message