# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Meeting Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Meeting {
  id: number;
  filename: string;
  original_filename: string;
  file_size: number;
  file_format: string;
  /** uploaded | transcribing | transcribed | failed */
  status: string;
  transcript?: string | null;
  transcribed_at?: string | null;
  uploaded_at: string;
}